remove conversion file
